requirements.txt
discord.py==1.3.4
python-dotenv==0.14.0
